added get contacts query
